requirements.txt
google-cloud-storage
google-cloud-firestore
google-cloud-bigquery
pandas
firebase-admin
pyspark==3.5.5
delta-spark==3.3.1
summa
nltk